Mapping fourth floor reliably, dropping virtual beacons.

+++ Final Challenge/ServoSPOT-OnCar/src/EC544/main/CarPoint.java
@@ -10,7 +10,7 @@ package EC544.main;
  */
 public class CarPoint {
 
-    public float time;
+    public double time;
     public float RF;
     public float LF;
     public float RR;
